Update research-statement.rst
watch for tense shifts (e.g. if one word CAN describe....it WOULD be "human"--> should be if one word COULD describe..it WOULD be human or if one word CAN describe...it IS human.

Can you use other words besides machine? It's used a lot in the first paragraph.